pyproject.toml
[project]
name = "kunai-sandbox"
version = "0.1.4"
description = "Simple malware sandbox powered by Qemu and Kunai"
readme = "README.md"
authors = [{ name = "qjerome", email = "quentin.jerome@circl.lu" }]
requires-python = ">=3.13"
dependencies = [
    "paramiko>=3.5.0",
    "pykunai>=0.1.5",
    "pyyaml>=6.0.2",
    "qemu-qmp>=0.0.3",
]

[project.scripts]
kunai-sandbox = "kunai_sandbox.main:main"
ks-gen-config = "kunai_sandbox.tools.gen_config:main"
ks-sandbox-init = "kunai_sandbox.tools.sandbox_init:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true
